Replace Signup function, changed minor outlook
For the signup button, previously it was made to redirect
to the web version of the signup page. Now it shows a dialog
popup for the information. However the actually functionality
is not implemented as we need to also include a email verification
which I plan to have another popup dialog box for that. Also added
a forgot password option to the login page where functionality is
also not implemented yet for the same reason.