Minor UI changes to pages, rating_page incomplete
Most of the stuff is done. Rating page still needs to have a function
to be implemented for the submit button (api call to push the score to
the database). Will most likely change the UI for the rating page as
well (current plan is to change background color and add images for
different score values). Still need to implement the verify email,
which is right after the signup.