Changes Recruiter page, signup function changed
The signup button no longer displays the signup dialog box but instead
now directs to a webview page of the website version. This was done due
to the need to stress that on signup people need to upload resume.
Recruiter page was changed so that each event now has its own scan
button. This way the event id will be passed to the rating page for it
to perform the necessary API call that needs a event id (api/scans).